Created a Home component
